+++ example/lib/hive/hive_adapters.g.dart
@@ -1,14 +1,46 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'user.dart';
+part of 'hive_adapters.dart';
 
 // **************************************************************************
-// TypeAdapterGenerator
+// AdaptersGenerator
 // **************************************************************************
+
+class ContactAdapter extends TypeAdapter<Contact> {
+  @override
+  final typeId = 1;
+
+  @override
+  Contact read(BinaryReader reader) {
+    final numOfFields = reader.readByte();
+    final fields = <int, dynamic>{
+      for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
+    };
+    return Contact(name: fields[0] as String);
+  }
+
+  @override
+  void write(BinaryWriter writer, Contact obj) {
+    writer
+      ..writeByte(1)
+      ..writeByte(0)
+      ..write(obj.name);
+  }
+
+  @override
+  int get hashCode => typeId.hashCode;
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is ContactAdapter &&
+          runtimeType == other.runtimeType &&
+          typeId == other.typeId;
+}
 
 class UserAdapter extends TypeAdapter<User> {
   @override
-  final int typeId = 3;
+  final typeId = 2;
 
   @override
   User read(BinaryReader reader) {
@@ -16,10 +48,11 @@ class UserAdapter extends TypeAdapter<User> {
     final fields = <int, dynamic>{
       for (int i = 0; i < numOfFields; i++) reader.readByte(): reader.read(),
     };
-    return User()
-      ..name = fields[0] as String
-      ..address = fields[1] as String
-      ..users = (fields[2] as List).cast<User>();
+    return User(
+      name: fields[0] as String,
+      address: fields[1] as String,
+      users: (fields[2] as List).cast<User>(),
+    );
   }
 
   @override